Replace f8e16034 fix with a more straightforward one
The purpose of the f8e16034 fix (Fix conditional location) was apparently
to protect against the fact that seqinfoForGenome() can return a NULL. But
the fix reorganizes the code in such a way that it's not clear anymore that
the warning issued in case of a disagreement between the sequence names in
the GTF or GFF file and the genome assembly specified via the 'genome'
argument will always be issued in the right context i.e. when 'genome' is
actually a single non-NA string.

The new fix protects against the fact that seqinfoForGenome() can return
a NULL by just starting the conditional with !is.null(ans_seqinfo) && ...
No need to reorganize the code.