track argument to ucscTableQuery() is defunct after long deprecation.
The UCSC table browser now requires an API key, so only the keyless
REST API is supported going forward.

This also makes the helper ucscTables() defunct, since it relied on
the table browser and its sole purpose was to help users transition to
using table identifiers instead of track identifiers.